requirements.txt
# Gemini API and Google Auth
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Text-to-Speech and Video Editing
gTTS
moviepy==1.0.3